add own step for queue setup
